pyproject.toml
[project]
name = "multilingual-rag-copilot"
version = "0.2.0"
description = "Multilingual RAG Copilot: chat with your documents with citations (deployable with a public URL)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.36",
  "pypdf>=4.2",
  "python-docx>=1.1",
  "beautifulsoup4>=4.12",
  "sentence-transformers>=3.0",
  "chromadb>=0.5.5",
  "groq>=0.11.0",
  "requests>=2.32",
  "mlflow>=2.14",
  "pandas>=2.1",
  "python-dotenv>=1.0",
  "passlib[bcrypt]>=1.7.4",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
